requirements.txt
mouseinfo==0.1.3 ; python_version >= "3.11" and python_version < "4.0"
pillow==10.2.0 ; python_version >= "3.11" and python_version < "4.0"
pyautogui==0.9.54 ; python_version >= "3.11" and python_version < "4.0"
pygetwindow==0.0.9 ; python_version >= "3.11" and python_version < "4.0"
pymsgbox==1.0.9 ; python_version >= "3.11" and python_version < "4.0"
pyobjc-core==10.1 ; python_version >= "3.11" and python_version < "4.0" and platform_system == "Darwin"
pyobjc-framework-cocoa==10.1 ; python_version >= "3.11" and python_version < "4.0" and platform_system == "Darwin"
pyobjc-framework-quartz==10.1 ; python_version >= "3.11" and python_version < "4.0" and platform_system == "Darwin"
pyperclip==1.8.2 ; python_version >= "3.11" and python_version < "4.0"
pyrect==0.2.0 ; python_version >= "3.11" and python_version < "4.0"
pyscreeze==0.1.30 ; python_version >= "3.11" and python_version < "4.0"
python3-xlib==0.15 ; platform_system == "Linux" and python_version >= "3.11" and python_version < "4.0"
pytweening==1.2.0 ; python_version >= "3.11" and python_version < "4.0"
rubicon-objc==0.4.7 ; python_version >= "3.11" and python_version < "4.0" and platform_system == "Darwin"
